lesson 117 authorized edit and delete added

+++ backend/routes/posts.js
@@ -63,10 +63,12 @@ router.put("/:id",checkAuth, multer({storage: storage}).single("image"), (req, r
      content: req.body.content,
      imagePath: imagePath
   });
-  console.log(post);
-   Post.updateOne({_id: req.params.id}, post).then(result => {
-     console.log(result);
-     res.status(200).json({ message: "Update Succesful"});
+   Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+     if (result.nModified > 0){
+      res.status(200).json({ message: "Update Succesful"});
+     } else {
+      res.status(401).json({ message: "Not Authorized"});
+     }
    })
  });
 
@@ -108,9 +110,13 @@ router.get("", (req,res,next) => {
 });
 
 router.delete("/:id",checkAuth, (req,res,next) => {
- Post.deleteOne({_id: req.params.id}).then(result => {
-  console.log(result);
-  res.status(200).json({messsage: "Post deleted"});
+ Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+  if (result.n > 0){
+    console.log(result);
+    res.status(200).json({ message: "Deletion Succesful"});
+   } else {
+    res.status(401).json({ message: "Deletion Not Authorized"});
+   }
  });
 });
